backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
openai
groq
anthropic
httpx>=0.25.0
websockets
cachetools
python-jose[cryptography]
tenacity
pydantic>=2.5.0
pydantic-settings
asyncio
python-dotenv>=1.0.0
